Added the movements for kope the NPC.
He has certain parameters. Firstly he selects the duration
for which he will walk, and then the direction in which he will
walk. After choosing both randomly he will switch to that direction.

Kope will roam around the map randomly and in irder to interact with
him, we need to find him and catch him.